+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### changed
 * avoid sending the same sequence of blocks multiple times
 * updated dependencies
+* removed `dirs` dependency
 
 ## [0.8.0] - 2023-04-22
 ### added